+++ src/i18n/en.json
@@ -88,8 +88,10 @@
     "error_loading_contract": "There has been an error while loading contract",
     "error_uploading_front_image": "Error uploading front picture",
     "error_uploading_back_image": "Error uploading back picture",
-    "error_while_saving_data": "There has been an error while saving the data"
-
+    "error_while_saving_data": "There has been an error while saving the data",
+    "GMFL-400": "Google map failed to load!",
+    "PNF-404": "Page not found!",
+    "ASB-502": "Already submitted your booking"
   },
   "title": {
     "feed_back_form": "Treatment Received Feedback Form"
